# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 衍图 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
  /skills/{id}/package:
    put:
      tags: [skills]
      summary: 上传已有 ZIP 技能的新版本
      description: 仅技能作者可上传。保留原技能 ID，并将 ZIP 解析为新的不可变技能包版本。
      operationId: updateZipSkillPackage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 已将新 ZIP 包追加为该技能的当前版本，旧版本及用户引用保留
        "400":
          description: 技能不存在、不是 ZIP 技能或上传包无效/超过大小限制
        "403":
          description: 当前用户不是技能作者
  /production-runs/{id}/steps/result:
    post:
      tags: [production]
      summary: 记录制作质检步骤结果
      description: |
        将实际执行的程序质检、语义质检或连续性检查证据写回已有的 check/verify 步骤。状态由 checks 推导，使用 expectedRevision 做并发控制。
        quality/continuity 步骤必须包含 semanticQuality；缺少或 unavailable 时记录 uncertain，不能将缺失的语义证据标成成功。
        证据限制为 64 KiB JSON，不能包含凭证字段或媒体 URL。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedRevision, stepId, evidence]
              properties:
                expectedRevision:
                  type: integer
                  minimum: 0
                stepId:
                  type: string
                  minLength: 1
                evidence:
                  type: object
                  required: [checks]
                  properties:
                    checks:
                      type: array
                      minItems: 1
                      maxItems: 200
                      items:
                        type: object
                        required: [name, status]
                        properties:
                          name:
                            type: string
                          status:
                            type: string
                            enum: [passed, failed, uncertain]
                    semanticQuality:
                      type: object
                      properties:
                        status:
                          type: string
                          enum: [passed, failed, uncertain, unavailable]
                  additionalProperties: true
      responses:
        "200":
          description: 已记录证据并返回更新后的 ProductionRun
        "400":
          description: 证据格式无效或超出限制
        "404":
          description: 制作运行不存在或不属于当前用户
        "409":
          description: run 版本已变化、步骤类型错误或依赖尚未完成
  /production-runs/{id}/steps/render:
    post:
      tags: [production]
      summary: 按 ProductionPlan 提交主时间线渲染
      description: |
        只接受当前 run 中 Kind=render 的有效步骤，并核对依赖、revision、重试 lineage、完整 manifest、分镜片段/音轨/字幕与最新成功生成任务的真实资源绑定。
        相同 idempotencyKey 和请求正文只返回原 attempt/task；不确定 attempt 不会重新排队。新 attempt 通过 retryOf 指向最近一次已结清 attempt。
        输出画幅和帧率必须符合已固化的 Brief 与 ProductionPlan 时间基准。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedRevision, stepId, idempotencyKey, timeline]
              properties:
                expectedRevision: { type: integer, minimum: 0 }
                stepId: { type: string, minLength: 1 }
                idempotencyKey: { type: string, minLength: 1, maxLength: 120 }
                retryOf: { type: string }
                timeline:
                  type: object
                  required: [projectId, timeline]
                  properties:
                    projectId: { type: string, minLength: 1 }
                    timeline:
                      type: object
                      description: TimelineProject；视频 clip 带 storyboardRowId/segmentId，音频和字幕 clip 带 storyboardRowId，并通过 directMedia.storageKey 引用对应生成任务的 resource:<id>
                      additionalProperties: true
                    output: { type: object, additionalProperties: true }
      responses:
        "200": { description: 返回持久化 render attempt 与 taskId；幂等重放返回同一任务 }
        "400": { description: 时间线或交付参数无效 }
        "404": { description: run 或资源不存在或无权访问 }
        "409": { description: revision、步骤依赖、绑定清单、画幅或幂等键冲突 }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
...